+++ Mariya/main.py
@@ -11,73 +11,22 @@ from report_module import ReportModule
 from requests_module import RequestsModule
 
 
-async def progress_bar(steps: int, max_lenght: int = 50):
-	step_size = 50 / steps
-	#step_size = 100 / max_lenght
-	for i in range(1, steps + 1):
-		print(f'\r{int(i * step_size)}%', end='')	#очень интересная тема
-		counter = int(i * step_size)
-		await asyncio.sleep(0.05)
-#		Ui_Dialog.on_progress_change(10)
-		return counter
-
-	#return('')										#очень интересная тема ^^^
-
-
-#async def request() -> None:
-#	request = RequestsModule()
-#	report = ReportModule()
-#	list_inn = report.read_exel_inn()
-#	for inn in list_inn:
-#		print(inn)
-#		data = request.report_generation(request.get_organization(request.get_id(inn)), inn)
-#		print(data)
-#		for organization in data:
-#			await report.writer_a_report_file(organization)
-#		print('Write >> OK')
-#		await asyncio.sleep(2)
-#	report.formater_to_exel()
-			
-
-def request() -> None:
-	bufered_data = []
-	report = ReportModule()
-	for inn in report.read_exel_inn():
-		request = RequestsModule(inn).get_organization()
-		for data in request:
-			bufered_data.append(data)
-	return bufered_data
-
-async def write_in_exel() -> None:
-	report = ReportModule()
-	for data in request():
-		report.writer_a_report_file(data)
-
-async def main_Function():
-	#try:
-		tasks = [
-			progress_bar(100, 100),
-	   		write_in_exel()
-		]
-		await asyncio.gather(*tasks)
-		
-	#except Exception as ex:
-	#	print(f'Ошибка: {ex}')
-
-
 class App(QMainWindow):
 	def __init__(self):
 		super(App, self).__init__()
 		self.ui = Ui_Dialog()
 		self.ui.setupUi(self)
-		#self.ui.progressBar.valueChanged(self.progress_bar)
+		#self.ui.progressBar.setValue()
+
 
 def Start_app():
-
+	
 	app = QApplication(sys.argv)
 	window = App()
 	window.show()
 	sys.exit(app.exec())	
+
+
 
 if __name__ == '__main__':
 #	start = time.time()
